Rename to HomescreenIllustrationRepository
